RDR-909 Added new datatype: email
This is mapped to django.forms.EmailField, so we get email validation for free.

To use it, just use email instead of string for a CDE's datatype.

Django EmailFields come with their own widget, so there is no need to specify a further custom widget name in the cde definition.