RDR-1324 Fix loading of jquery
Web browser was blocking the CDN because the URL was not HTTPS.

In future we should be using bower or npm to fetch vendor javascript,
and concatenating and minifying before serving in production.

CDN is not suitable because an unknown 3rd party is given access to run
any script within registry pages. Also mixing in three different CDNs
increases the chance of something not working.