Adds possibility to order follow up forms by their name.
The follow up forms are currently ordered by creation time.

This feature allows the registry owners to change the ordering
of the links to the follow up forms by using:

Admin -> Registry Context Form Groups

The ordering can be set to "Creation Time", which is the current
behaviour, or to "Name" which will order the links by their name, which
for FH is the Date of Assessment.