fix: #819 abnormality_condition support for [ and =="str"
1) abnormality condition supported x in ( ) instead or x in [ ]
Tuples were not the best solution as it required single , at the end for unique values.

2) x == "string" were not supported.
   PS: we do not support x === "string"

3) In designer mode, when the designer set the abnormality condition rules,
we now validate them. Note that we kept the validation during a form update. So if
we change the rule validation in the code but forget to update
rules in the DB, we will know about these newly wrong rules
(.i.e. the form values won't be checked against rules that were correct previously but not anymore).

4) Finally as the validation functions do not require any access to the instance
values, we make them static.