issue #897 - update_calculated_fields crashes on unexisting key
Problem: the django admin update_Calculated_fields is crashing on staging as
it is calling save_form_value which load all dynamic data and save them all.
The problem been that some key (form__section_cdecode) do not exist, the form name
having changed. When saving the dynamic data, it crashes as the key does not exist.
To solve the issue we add a new parameter ignoring the not existing keys.